Add start browser test. (#44)
